AST-to-CFG phase without breaks (+ proof)
Other changes in this commit:
* Proof of normalization phase that groups together simple commands before
  transforming the AST to a CFG (in order to reduce the number of blocks).
* Adjust scope semantics to also work if there are clashes between
scoped declarations.
* Adjust CFG semantics to not take acyclicity as a precondition but to instead
do a case distinction on acyclicity (in the case where the CFG is not acyclic,
None is returned). This makes it easier to work with during the AST-to-CFG phase,
since one need not guarantee the precondition at all points and in certain cases
need not reprove acyclicity of parts of the CFG.
* Restructure SimpleCmd to have a separate sequential composition.
This makes it cleaner to deal with CFGs where a BasicBlock is now expressed
as a SimpleCmd instead of seq<SimpleCmd>.
* The deep embedding of the weakest precondition was removed, since at the
AST level we probably do not need it (and the shallow embedding is
easier to work with).