First version passification implementation (#1)
First implementation for passification (desugaring of assignments to assumptions via SSA)

This commit comments out quantifiers to be able to define a first version of substitution that does not need to deal with variable capture. Support for quantifiers should be added by using a common binder structure (e.g., DeBruijn indices or a locally nameless representation). This commit also comments out old expressions, because they are currently not given a semantics (they were just part of the syntax).

In the SSA implementation, there is one point where the implementation needs to iterate over a set. This is achieved by transforming the set to a sequence in a deterministic way by first populating an array (nondeterministically) and then sorting the  array. In general, this is more efficient than choosing the next element in the iteration to be the smallest element, but still it would be nice to have a more efficient version that does not require sorting.

The sorting function is currently not given an implementation (should update this soon).